Refacto - MovingEntities as PacMan and Ghost common parent
