+++ src/main/java/fr/LaurentFE/pacManClone/entities/Ghost.java
@@ -1,6 +1,5 @@
 package fr.LaurentFE.pacManClone.entities;
 
-import fr.LaurentFE.pacManClone.*;
 import fr.LaurentFE.pacManClone.entities.ghostPersonality.GhostPersonality;
 import fr.LaurentFE.pacManClone.map.GameMap;
 import fr.LaurentFE.pacManClone.map.Position;
@@ -12,16 +11,12 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Random;
 
-public class Ghost {
-    private Orientation orientation;
-    private final int moveSpeed;
-    private final Rectangle hitBox;
+public class Ghost extends MovingEntity {
     private final Color color;
     private GhostState state;
     private final GhostPersonality chasePersonality;
     private final TileIndex scatterTargetTile;
     private  final TileIndex eatenTargetTile = GameMap.getInstance().getGhostHouse();
-    private final GameMap gameMap;
     private TileIndex lastCrossroadTile;
     private long chaseNanoTimeStart;
     private long scatterNanoTimeStart;
@@ -37,14 +32,11 @@ public class Ghost {
                  Color color,
                  GhostPersonality chasePersonality,
                  TileIndex scatterTargetTile) {
-        hitBox = new Rectangle(startingPosition.x, startingPosition.y, GamePanel.TILE_SIZE, GamePanel.TILE_SIZE);
-        orientation = startingOrientation;
-        this.moveSpeed = moveSpeed;
+        super(startingPosition, startingOrientation, moveSpeed);
         this.color = color;
         state = GhostState.CHASE;
         this.chasePersonality = chasePersonality;
         this.scatterTargetTile = scatterTargetTile;
-        gameMap = GameMap.getInstance();
         lastCrossroadTile = new TileIndex(0, 0);
         chaseNanoTimeStart = System.nanoTime();
         scatterNanoTimeStart = 0;
@@ -52,24 +44,8 @@ public class Ghost {
         score = 100;
     }
 
-    public Rectangle getHitBox() {
-        return hitBox;
-    }
-
-    public Orientation getOrientation() {
-        return orientation;
-    }
-
-    public void setOrientation(Orientation orientation) {
-        this.orientation = orientation;
-    }
-
     public Color getColor() {
         return color;
-    }
-
-    public Position getPosition() {
-        return new Position(hitBox.x, hitBox.y);
     }
 
     public GhostState getState() {
@@ -118,124 +94,6 @@ public class Ghost {
 
     public TileIndex getLastCrossroadTile() {
         return lastCrossroadTile;
-    }
-
-    public void move() {
-        if (orientation == Orientation.LEFT) {
-            hitBox.x = hitBox.x - moveSpeed;
-            if (hitBox.x < -GamePanel.TILE_SIZE)
-                hitBox.x = (gameMap.getMapWidthTile() + 1) * GamePanel.TILE_SIZE + hitBox.x;
-        } else if (orientation == Orientation.RIGHT) {
-            hitBox.x = hitBox.x + moveSpeed;
-            if (hitBox.x >= gameMap.getMapWidthTile() * GamePanel.TILE_SIZE)
-                hitBox.x = hitBox.x - (gameMap.getMapWidthTile() + 1) * GamePanel.TILE_SIZE;
-        } else if (orientation == Orientation.UP) {
-            hitBox.y = hitBox.y - moveSpeed;
-        } else if (orientation == Orientation.DOWN) {
-            hitBox.y = hitBox.y + moveSpeed;
-        }
-    }
-
-    public void bumpOutOfCollision(TileIndex collisionTile) {
-        Position collisionPosition = collisionTile.toPosition();
-        if (orientation == Orientation.LEFT) {
-            hitBox.x = collisionPosition.x + GamePanel.TILE_SIZE;
-        } else if (orientation == Orientation.RIGHT) {
-            hitBox.x = collisionPosition.x - GamePanel.TILE_SIZE;
-        } else if (orientation == Orientation.UP) {
-            hitBox.y = collisionPosition.y + GamePanel.TILE_SIZE;
-        } else if (orientation == Orientation.DOWN) {
-            hitBox.y = collisionPosition.y - GamePanel.TILE_SIZE;
-        }
-    }
-
-    private void tileLoopAroundHorizontal(TileIndex tile) {
-        if (tile.x < 0)
-            tile.x = gameMap.getMapWidthTile() - 1;
-        if (tile.x >= gameMap.getMapWidthTile())
-            tile.x = 0;
-    }
-
-    private Rectangle getNextPathTileForOrientation(Orientation nextOrientation) {
-        TileIndex directionModifier;
-        if (nextOrientation == Orientation.UP) {
-            directionModifier = new TileIndex(0, -1);
-        } else if (nextOrientation == Orientation.LEFT) {
-            directionModifier = new TileIndex(-1, 0);
-        } else if (nextOrientation == Orientation.DOWN) {
-            directionModifier = new TileIndex(0, 1);
-        } else {
-            directionModifier = new TileIndex(1, 0);
-        }
-        TileIndex tileAIndex = new Position(hitBox.x, hitBox.y).toTileIndex()
-                .add(directionModifier);
-        TileIndex tileBIndex = new Position(hitBox.x, hitBox.y).toTileIndex()
-                .add(directionModifier);
-        TileIndex tileCIndex = new Position(hitBox.x, hitBox.y).toTileIndex()
-                .add(directionModifier);
-
-        if (nextOrientation == Orientation.UP || nextOrientation == Orientation.DOWN) {
-            tileAIndex.x -= 1;
-            tileCIndex.x += 1;
-        } else {
-            tileAIndex.y -= 1;
-            tileCIndex.y += 1;
-        }
-
-        tileLoopAroundHorizontal(tileAIndex);
-        tileLoopAroundHorizontal(tileBIndex);
-        tileLoopAroundHorizontal(tileCIndex);
-
-        Position tileAPosition = tileAIndex.toPosition();
-        Position tileBPosition = tileBIndex.toPosition();
-        Position tileCPosition = tileCIndex.toPosition();
-
-        if (canGoThroughTile(tileBIndex)) {
-            return new Rectangle(
-                    tileBPosition.x,
-                    tileBPosition.y,
-                    GamePanel.TILE_SIZE,
-                    GamePanel.TILE_SIZE
-            );
-        } else if (canGoThroughTile(tileAIndex)) {
-            return new Rectangle(
-                    tileAPosition.x,
-                    tileAPosition.y,
-                    GamePanel.TILE_SIZE,
-                    GamePanel.TILE_SIZE
-            );
-        } else if (canGoThroughTile(tileCIndex)) {
-            return new Rectangle(
-                    tileCPosition.x,
-                    tileCPosition.y,
-                    GamePanel.TILE_SIZE,
-                    GamePanel.TILE_SIZE
-            );
-        } else {
-            return new Rectangle();
-        }
-    }
-
-    private boolean canGetIntoPath(Orientation nextOrientation) {
-        Rectangle pathTile = getNextPathTileForOrientation(nextOrientation);
-        if (pathTile.equals(new Rectangle()))
-            return false;
-
-        if (nextOrientation == Orientation.UP
-                || nextOrientation == Orientation.DOWN) {
-            if (pathTile.x - hitBox.x < moveSpeed
-                    && pathTile.x - hitBox.x > -moveSpeed) {
-                hitBox.x = pathTile.x;
-                return true;
-            }
-        } else {
-            if (pathTile.y - hitBox.y < moveSpeed
-                    && pathTile.y - hitBox.y > -moveSpeed) {
-                hitBox.y = pathTile.y;
-                return true;
-            }
-        }
-        return false;
     }
 
     private boolean isCurrentTileACrossroad() {
@@ -415,7 +273,7 @@ public class Ghost {
         if (canGetIntoPath(nextOrientation)) {
             if (isCurrentTileACrossroad())
                 lastCrossroadTile = new Position(hitBox.x, hitBox.y).toTileIndex();
-            setOrientation(nextOrientation);
+            orientation = nextOrientation;
             updatePosition();
             return true;
         }
@@ -445,29 +303,6 @@ public class Ghost {
             chaseNanoTimeStart = currentTime;
             orientation = Orientation.UP;
             lastCrossroadTile = new TileIndex(0,0);
-        }
-    }
-
-    private void updatePosition() {
-        move();
-        TileIndex upperLeftTile = new Position(hitBox.x, hitBox.y).toTileIndex();
-        TileIndex upperRightTile = new Position(hitBox.x + hitBox.width-1, hitBox.y).toTileIndex();
-        TileIndex lowerLeftTile = new Position(hitBox.x, hitBox.y + hitBox.height-1).toTileIndex();
-        TileIndex lowerRightTile = new Position(hitBox.x + hitBox.width-1, hitBox.y + hitBox.height-1).toTileIndex();
-
-        tileLoopAroundHorizontal(upperLeftTile);
-        tileLoopAroundHorizontal(upperRightTile);
-        tileLoopAroundHorizontal(lowerLeftTile);
-        tileLoopAroundHorizontal(lowerRightTile);
-
-        if (!canGoThroughTile(upperLeftTile)) {
-            bumpOutOfCollision(upperLeftTile);
-        } else if (!canGoThroughTile(upperRightTile)) {
-            bumpOutOfCollision(upperRightTile);
-        } else if (!canGoThroughTile(lowerLeftTile)) {
-            bumpOutOfCollision(lowerLeftTile);
-        } else if (!canGoThroughTile(lowerRightTile)) {
-            bumpOutOfCollision(lowerRightTile);
         }
     }
