+++ src/main/java/fr/LaurentFE/pacManClone/entities/PacMan.java
@@ -1,44 +1,29 @@
 package fr.LaurentFE.pacManClone.entities;
 
-import fr.LaurentFE.pacManClone.*;
-import fr.LaurentFE.pacManClone.map.GameMap;
 import fr.LaurentFE.pacManClone.map.Position;
 import fr.LaurentFE.pacManClone.map.TileIndex;
 import fr.LaurentFE.pacManClone.map.TileType;
 
-import java.awt.*;
-
-public class PacMan {
-    private Orientation orientation;
+public class PacMan extends MovingEntity {
     private final int maxMouthAngle;
     private int currentMouthAngle;
     private int mouthAngleIncrement;
     private boolean isAlive;
     private int lives;
     private boolean deathAnimationFinished;
-    private final int moveSpeed;
-    private final Rectangle hitBox;
-    private final GameMap gameMap;
 
     public PacMan(Position startingPosition, Orientation startingOrientation, int moveSpeed, int lives) {
-        hitBox = new Rectangle(startingPosition.x, startingPosition.y, GamePanel.TILE_SIZE, GamePanel.TILE_SIZE);
-        orientation = startingOrientation;
-        this.moveSpeed = moveSpeed;
+        super(startingPosition, startingOrientation, moveSpeed);
         maxMouthAngle = 90;
         currentMouthAngle = maxMouthAngle;
         mouthAngleIncrement = -5;
         isAlive = true;
         this.lives = lives;
         deathAnimationFinished = false;
-        gameMap = GameMap.getInstance();
     }
 
     public boolean isDeathAnimationFinished() {
         return deathAnimationFinished;
-    }
-
-    public Rectangle getHitBox() {
-        return hitBox;
     }
 
     public int getLives() {
@@ -60,159 +45,14 @@ public class PacMan {
         }
     }
 
-    public void move() {
-            if (orientation == Orientation.LEFT) {
-                hitBox.x = hitBox.x - moveSpeed;
-                if (hitBox.x < -GamePanel.TILE_SIZE)
-                    hitBox.x = (gameMap.getMapWidthTile() + 1) * GamePanel.TILE_SIZE + hitBox.x;
-            } else if (orientation == Orientation.RIGHT) {
-                hitBox.x = hitBox.x + moveSpeed;
-                if (hitBox.x >= gameMap.getMapWidthTile() * GamePanel.TILE_SIZE)
-                    hitBox.x = hitBox.x - (gameMap.getMapWidthTile() + 1) * GamePanel.TILE_SIZE;
-            } else if (orientation == Orientation.UP) {
-                hitBox.y = hitBox.y - moveSpeed;
-            } else if (orientation == Orientation.DOWN) {
-                hitBox.y = hitBox.y + moveSpeed;
-            }
-    }
-
-    public void bumpOutOfCollision(TileIndex collisionTileIndex) {
-        Position collisionTilePosition = collisionTileIndex.toPosition();
-        if (orientation == Orientation.LEFT) {
-            hitBox.x = collisionTilePosition.x + GamePanel.TILE_SIZE;
-        } else if (orientation == Orientation.RIGHT) {
-            hitBox.x = collisionTilePosition.x - GamePanel.TILE_SIZE;
-        } else if (orientation == Orientation.UP) {
-            hitBox.y = collisionTilePosition.y + GamePanel.TILE_SIZE;
-        } else if (orientation == Orientation.DOWN) {
-            hitBox.y = collisionTilePosition.y - GamePanel.TILE_SIZE;
-        }
-    }
-
-    public Orientation getOrientation() {
-        return orientation;
-    }
-
     public int getCurrentMouthAngle() {
         return currentMouthAngle;
     }
 
-    public Position getPosition() {
-        return new Position(hitBox.x, hitBox.y);
-    }
-
-    private void tileLoopAroundHorizontal(TileIndex tile) {
-        if (tile.x < 0)
-            tile.x = gameMap.getMapWidthTile() - 1;
-        if (tile.x >= gameMap.getMapWidthTile())
-            tile.x = 0;
-    }
-
-    private Rectangle getNextPathTileForOrientation(Orientation nextOrientation) {
-        TileIndex directionModifier;
-        if (nextOrientation == Orientation.UP) {
-            directionModifier = new TileIndex(0, -1);
-        } else if (nextOrientation == Orientation.LEFT) {
-            directionModifier = new TileIndex(-1, 0);
-        } else if (nextOrientation == Orientation.DOWN) {
-            directionModifier = new TileIndex(0, 1);
-        } else {
-            directionModifier = new TileIndex(1, 0);
-        }
-        TileIndex tileAIndex = new Position(hitBox.x, hitBox.y).toTileIndex()
-                .add(directionModifier);
-        TileIndex tileBIndex = new Position(hitBox.x, hitBox.y).toTileIndex()
-                .add(directionModifier);
-        TileIndex tileCIndex = new Position(hitBox.x, hitBox.y).toTileIndex()
-                .add(directionModifier);
-
-        if (nextOrientation == Orientation.UP || nextOrientation == Orientation.DOWN) {
-            tileAIndex.x -= 1;
-            tileCIndex.x += 1;
-        } else {
-            tileAIndex.y -= 1;
-            tileCIndex.y += 1;
-        }
-
-        tileLoopAroundHorizontal(tileAIndex);
-        tileLoopAroundHorizontal(tileBIndex);
-        tileLoopAroundHorizontal(tileCIndex);
-
-        Position tileAPosition = tileAIndex.toPosition();
-        Position tileBPosition = tileBIndex.toPosition();
-        Position tileCPosition = tileCIndex.toPosition();
-
-        if (canGoThroughTile(tileBIndex)) {
-            return new Rectangle(
-                    tileBPosition.x,
-                    tileBPosition.y,
-                    GamePanel.TILE_SIZE,
-                    GamePanel.TILE_SIZE
-            );
-        } else if (canGoThroughTile(tileAIndex)) {
-            return new Rectangle(
-                    tileAPosition.x,
-                    tileAPosition.y,
-                    GamePanel.TILE_SIZE,
-                    GamePanel.TILE_SIZE
-            );
-        } else if (canGoThroughTile(tileCIndex)) {
-            return new Rectangle(
-                    tileCPosition.x,
-                    tileCPosition.y,
-                    GamePanel.TILE_SIZE,
-                    GamePanel.TILE_SIZE
-            );
-        } else {
-            return new Rectangle();
-        }
-    }
-
-    private boolean canGetIntoPath(Orientation nextOrientation) {
-        Rectangle pathTile = getNextPathTileForOrientation(nextOrientation);
-        if (pathTile.equals(new Rectangle()))
-            return false;
-
-        if (nextOrientation == Orientation.UP
-                || nextOrientation == Orientation.DOWN) {
-            if (pathTile.x - hitBox.x < moveSpeed
-                    && pathTile.x - hitBox.x > -moveSpeed) {
-                hitBox.x = pathTile.x;
-                return true;
-            }
-        } else {
-            if (pathTile.y - hitBox.y < moveSpeed
-                    && pathTile.y - hitBox.y > -moveSpeed) {
-                hitBox.y =pathTile.y;
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private void updatePosition() {
+    protected void updatePosition() {
         move();
-        TileIndex upperLeftTile = new Position(hitBox.x,hitBox.y).toTileIndex();
-        TileIndex upperRightTile = new Position(hitBox.x + hitBox.width-1, hitBox.y).toTileIndex();
-        TileIndex lowerLeftTile = new Position(hitBox.x, hitBox.y + hitBox.height-1).toTileIndex();
-        TileIndex lowerRightTile = new Position(hitBox.x + hitBox.width-1,hitBox.y + hitBox.height-1).toTileIndex();
-
-        tileLoopAroundHorizontal(upperLeftTile);
-        tileLoopAroundHorizontal(upperRightTile);
-        tileLoopAroundHorizontal(lowerLeftTile);
-        tileLoopAroundHorizontal(lowerRightTile);
-
-        if (!canGoThroughTile(upperLeftTile)) {
-            bumpOutOfCollision(upperLeftTile);
-        } else if (!canGoThroughTile(upperRightTile)) {
-            bumpOutOfCollision(upperRightTile);
-        } else if (!canGoThroughTile(lowerLeftTile)) {
-            bumpOutOfCollision(lowerLeftTile);
-        } else if (!canGoThroughTile(lowerRightTile)) {
-            bumpOutOfCollision(lowerRightTile);
-        } else {
+        if (!checkForWallCollisions())
             animateMouth();
-        }
     }
 
     public boolean canGoThroughTile(TileIndex tileIndex) {
